requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asyncpg==0.27.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53
discord-ext-menus-views @ git+https://github.com/oliver-ni/discord-ext-menus-views@a472f3de648eb9e0fcffc2e101c69841de92f3ff
discord.py==2.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
requests-oauthlib==1.3.1
six==1.16.0
toml==0.10.2
typing_extensions==4.5.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
yarl==1.8.2
